Refactor ia2.h into public and private headers
Adds ia2_internal.h for internal implementation definitions that users
don't need to see. Refactors everything internal into this header and
combines macro definitions.
