+++ libia2/include/permissive_mode.h
@@ -13,6 +13,7 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <string.h>
+#include <sys/ucontext.h>
 
 #include "ia2.h"
